move fragments and viewModels to presentation layer.
