+++ app/src/main/java/com/nima/tmdb/framewrok/presentation/list/MovieListViewModel.kt
@@ -1,4 +1,4 @@
-package com.nima.tmdb.viewModels
+package com.nima.tmdb.framewrok.presentation.list
 
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
