+++ app/src/main/java/com/nima/tmdb/framewrok/presentation/login/LoginViewModel.kt
@@ -1,26 +1,21 @@
-package com.nima.tmdb.viewModels
+package com.nima.tmdb.framewrok.presentation.login
 
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.nima.tmdb.business.domain.model.login.*
-import com.nima.tmdb.business.interactors.splash.SplashInteractors
-import com.nima.tmdb.database.entities.UserInfo
+import com.nima.tmdb.business.domain.model.login.LoginInfo
+import com.nima.tmdb.business.domain.model.login.LoginResponse
+import com.nima.tmdb.business.domain.model.login.RequestToken
+import com.nima.tmdb.business.domain.model.login.Session
+import com.nima.tmdb.business.interactors.login.LoginInteractors
 import com.nima.tmdb.requests.wrapper.ApiWrapper
 import kotlinx.coroutines.launch
 
-class SplashViewModel @ViewModelInject constructor(
-    private val repository: SplashInteractors
+class LoginViewModel @ViewModelInject constructor(
+    private val repository: LoginInteractors
 ) : ViewModel() {
-    private val _token = MutableLiveData<ApiWrapper<Token>>()
-    val getToken: LiveData<ApiWrapper<Token>>
-        get() = _token
-
-    private val _userInfo = MutableLiveData<UserInfo>()
-    val userInfo: LiveData<UserInfo>
-        get() = _userInfo
 
     private val _login = MutableLiveData<ApiWrapper<LoginResponse>>()
     val login: LiveData<ApiWrapper<LoginResponse>>
@@ -29,11 +24,6 @@ class SplashViewModel @ViewModelInject constructor(
     private val _sessionId = MutableLiveData<ApiWrapper<Session>>()
     val sessionId: LiveData<ApiWrapper<Session>>
         get() = _sessionId
-
-    fun getToken(apiKey: String) =
-        viewModelScope.launch {
-            _token.value = repository.token.getToken(apiKey)
-        }
 
     fun login(loginInfo: LoginInfo, apiKey: String) =
         viewModelScope.launch {
