+++ app/src/main/res/layout/fragment_main_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.fragments.MainPageFragment">
+    tools:context=".framewrok.presentation.mainPage.MainPageFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
