+++ app/src/main/res/layout/fragment_splash.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.fragments.SplashFragment">
+    tools:context=".framewrok.presentation.splash.SplashFragment">
 
     <Button
         android:id="@+id/btn_firstPageF_animation"
